requirements.txt
streamlit
python_dotenv == 1.0.1
langchain
pinecone-client
crewai == 0.51.1
crewai-tools == 0.12.0

PyPDF2 == 3.0.1